requirements.txt
matplotlib==3.4.2
numpy==1.21.6
pandas==1.1.0
Pillow==8.4.0
scikit_learn==1.0.2
torch==1.11.0+cu113
torchvision==0.12.0+cu113
python == 3.7.1